refactor(editor): extract vscode settings path and editor config into new type files
