+++ src/input/editorConfiguration.ts
@@ -1,0 +1,1 @@
+export type EditorConfiguration = { [key: string]: any };
