+++ src/input/vsCodeSettings.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_VSCODE_SETTINGS_PATH = ".vscode/settings.json";
